fix: Improve Terraform agent<->resource association testing
There were a few bugs with multiple agent association this fixes,
and it adds thorough tests using real Terraform state and plans.
