+++ .vscode/settings.json
@@ -38,6 +38,7 @@
     "kirsle",
     "ldflags",
     "manifoldco",
+    "mapstructure",
     "mattn",
     "mitchellh",
     "moby",
@@ -67,9 +68,11 @@
     "tcpip",
     "TCSETS",
     "templateversions",
+    "testdata",
     "testid",
     "tfexec",
     "tfjson",
+    "tfplan",
     "tfstate",
     "trimprefix",
     "turnconn",
