+++ provisioner/terraform/provision.go
@@ -13,10 +13,8 @@ import (
 	"runtime"
 	"strings"
 
-	"github.com/awalterschulze/gographviz"
 	"github.com/hashicorp/terraform-exec/tfexec"
 	tfjson "github.com/hashicorp/terraform-json"
-	"github.com/mitchellh/mapstructure"
 	"golang.org/x/xerrors"
 
 	"github.com/coder/coder/provisionersdk"
@@ -25,7 +23,7 @@ import (
 
 var (
 	// noStateRegex is matched against the output from `terraform state show`
-	noStateRegex = regexp.MustCompile(`(?i)State read error.*no state`)
+	noStateRegex = regexp.MustCompile(`no state`)
 )
 
 // Provision executes `terraform apply`.
@@ -202,7 +200,7 @@ func (t *terraform) Provision(stream proto.DRPCProvisioner_ProvisionStream) erro
 	// contingency, in the future we will try harder to prevent workspaces being
 	// broken this hard.
 	if start.Metadata.WorkspaceTransition == proto.WorkspaceTransition_DESTROY {
-		_, err := getTerraformState(shutdown, terraform, statefilePath)
+		_, err := pullTerraformState(shutdown, terraform, statefilePath)
 		if xerrors.Is(err, os.ErrNotExist) {
 			_ = stream.Send(&proto.Provision_Response{
 				Type: &proto.Provision_Response_Log{
@@ -325,86 +323,9 @@ func parseTerraformPlan(ctx context.Context, terraform *tfexec.Terraform, planfi
 	if err != nil {
 		return nil, xerrors.Errorf("graph: %w", err)
 	}
-	resourceDependencies, err := findDirectDependencies(rawGraph)
+	resources, err := ConvertResources(plan.PlannedValues.RootModule, rawGraph)
 	if err != nil {
-		return nil, xerrors.Errorf("find dependencies: %w", err)
-	}
-
-	resources := make([]*proto.Resource, 0)
-	agents := map[string]*proto.Agent{}
-
-	tfResources := make([]*tfjson.ConfigResource, 0)
-	var appendResources func(mod *tfjson.ConfigModule)
-	appendResources = func(mod *tfjson.ConfigModule) {
-		for _, module := range mod.ModuleCalls {
-			appendResources(module.Module)
-		}
-		tfResources = append(tfResources, mod.Resources...)
-	}
-	appendResources(plan.Config.RootModule)
-
-	// Store all agents inside the maps!
-	for _, resource := range tfResources {
-		if resource.Type != "coder_agent" {
-			continue
-		}
-		agent := &proto.Agent{
-			Name: resource.Name,
-			Auth: &proto.Agent_Token{},
-		}
-		if operatingSystemRaw, has := resource.Expressions["os"]; has {
-			operatingSystem, ok := operatingSystemRaw.ConstantValue.(string)
-			if ok {
-				agent.OperatingSystem = operatingSystem
-			}
-		}
-		if archRaw, has := resource.Expressions["arch"]; has {
-			arch, ok := archRaw.ConstantValue.(string)
-			if ok {
-				agent.Architecture = arch
-			}
-		}
-		if envRaw, has := resource.Expressions["env"]; has {
-			env, ok := envRaw.ConstantValue.(map[string]interface{})
-			if ok {
-				agent.Env = map[string]string{}
-				for key, valueRaw := range env {
-					value, valid := valueRaw.(string)
-					if !valid {
-						continue
-					}
-					agent.Env[key] = value
-				}
-			}
-		}
-		if startupScriptRaw, has := resource.Expressions["startup_script"]; has {
-			startupScript, ok := startupScriptRaw.ConstantValue.(string)
-			if ok {
-				agent.StartupScript = startupScript
-			}
-		}
-		if directoryRaw, has := resource.Expressions["dir"]; has {
-			dir, ok := directoryRaw.ConstantValue.(string)
-			if ok {
-				agent.Directory = dir
-			}
-		}
-
-		agents[convertAddressToLabel(resource.Address)] = agent
-	}
-
-	for _, resource := range tfResources {
-		if resource.Mode == tfjson.DataResourceMode {
-			continue
-		}
-		if resource.Type == "coder_agent" || resource.Type == "coder_agent_instance" || resource.Type == "coder_app" {
-			continue
-		}
-		resources = append(resources, &proto.Resource{
-			Name:   resource.Name,
-			Type:   resource.Type,
-			Agents: findAgents(resourceDependencies, agents, convertAddressToLabel(resource.Address)),
-		})
+		return nil, err
 	}
 
 	return &proto.Provision_Response{
@@ -420,185 +341,19 @@ func parseTerraformApply(ctx context.Context, terraform *tfexec.Terraform, state
 	_, err := os.Stat(statefilePath)
 	statefileExisted := err == nil
 
-	state, err := getTerraformState(ctx, terraform, statefilePath)
+	state, err := pullTerraformState(ctx, terraform, statefilePath)
 	if err != nil {
 		return nil, xerrors.Errorf("get terraform state: %w", err)
 	}
-
-	resources := make([]*proto.Resource, 0)
+	rawGraph, err := terraform.Graph(ctx)
+	if err != nil {
+		return nil, xerrors.Errorf("get terraform graph: %w", err)
+	}
+	var resources []*proto.Resource
 	if state.Values != nil {
-		rawGraph, err := terraform.Graph(ctx)
+		resources, err = ConvertResources(state.Values.RootModule, rawGraph)
 		if err != nil {
-			return nil, xerrors.Errorf("graph: %w", err)
-		}
-		resourceDependencies, err := findDirectDependencies(rawGraph)
-		if err != nil {
-			return nil, xerrors.Errorf("find dependencies: %w", err)
-		}
-		type agentAttributes struct {
-			Auth            string            `mapstructure:"auth"`
-			OperatingSystem string            `mapstructure:"os"`
-			Architecture    string            `mapstructure:"arch"`
-			Directory       string            `mapstructure:"dir"`
-			ID              string            `mapstructure:"id"`
-			Token           string            `mapstructure:"token"`
-			Env             map[string]string `mapstructure:"env"`
-			StartupScript   string            `mapstructure:"startup_script"`
-		}
-		agents := map[string]*proto.Agent{}
-
-		tfResources := make([]*tfjson.StateResource, 0)
-		var appendResources func(resource *tfjson.StateModule)
-		appendResources = func(mod *tfjson.StateModule) {
-			for _, module := range mod.ChildModules {
-				appendResources(module)
-			}
-			tfResources = append(tfResources, mod.Resources...)
-		}
-		appendResources(state.Values.RootModule)
-
-		// Store all agents inside the maps!
-		for _, resource := range tfResources {
-			if resource.Type != "coder_agent" {
-				continue
-			}
-			var attrs agentAttributes
-			err = mapstructure.Decode(resource.AttributeValues, &attrs)
-			if err != nil {
-				return nil, xerrors.Errorf("decode agent attributes: %w", err)
-			}
-			agent := &proto.Agent{
-				Name:            resource.Name,
-				Id:              attrs.ID,
-				Env:             attrs.Env,
-				StartupScript:   attrs.StartupScript,
-				OperatingSystem: attrs.OperatingSystem,
-				Architecture:    attrs.Architecture,
-				Directory:       attrs.Directory,
-			}
-			switch attrs.Auth {
-			case "token":
-				agent.Auth = &proto.Agent_Token{
-					Token: attrs.Token,
-				}
-			default:
-				agent.Auth = &proto.Agent_InstanceId{}
-			}
-			agents[convertAddressToLabel(resource.Address)] = agent
-		}
-
-		// Manually associate agents with instance IDs.
-		for _, resource := range tfResources {
-			if resource.Type != "coder_agent_instance" {
-				continue
-			}
-			agentIDRaw, valid := resource.AttributeValues["agent_id"]
-			if !valid {
-				continue
-			}
-			agentID, valid := agentIDRaw.(string)
-			if !valid {
-				continue
-			}
-			instanceIDRaw, valid := resource.AttributeValues["instance_id"]
-			if !valid {
-				continue
-			}
-			instanceID, valid := instanceIDRaw.(string)
-			if !valid {
-				continue
-			}
-
-			for _, agent := range agents {
-				if agent.Id != agentID {
-					continue
-				}
-				agent.Auth = &proto.Agent_InstanceId{
-					InstanceId: instanceID,
-				}
-				break
-			}
-		}
-
-		type appAttributes struct {
-			AgentID      string `mapstructure:"agent_id"`
-			Name         string `mapstructure:"name"`
-			Icon         string `mapstructure:"icon"`
-			URL          string `mapstructure:"url"`
-			Command      string `mapstructure:"command"`
-			RelativePath bool   `mapstructure:"relative_path"`
-		}
-		// Associate Apps with agents.
-		for _, resource := range state.Values.RootModule.Resources {
-			if resource.Type != "coder_app" {
-				continue
-			}
-			var attrs appAttributes
-			err = mapstructure.Decode(resource.AttributeValues, &attrs)
-			if err != nil {
-				return nil, xerrors.Errorf("decode app attributes: %w", err)
-			}
-			if attrs.Name == "" {
-				// Default to the resource name if none is set!
-				attrs.Name = resource.Name
-			}
-			for _, agent := range agents {
-				if agent.Id != attrs.AgentID {
-					continue
-				}
-				agent.Apps = append(agent.Apps, &proto.App{
-					Name:         attrs.Name,
-					Command:      attrs.Command,
-					Url:          attrs.URL,
-					Icon:         attrs.Icon,
-					RelativePath: attrs.RelativePath,
-				})
-			}
-		}
-
-		for _, resource := range tfResources {
-			if resource.Mode == tfjson.DataResourceMode {
-				continue
-			}
-			if resource.Type == "coder_agent" || resource.Type == "coder_agent_instance" || resource.Type == "coder_app" {
-				continue
-			}
-			resourceAgents := findAgents(resourceDependencies, agents, convertAddressToLabel(resource.Address))
-			for _, agent := range resourceAgents {
-				// Didn't use instance identity.
-				if agent.GetToken() != "" {
-					continue
-				}
-
-				key, isValid := map[string]string{
-					"google_compute_instance":         "instance_id",
-					"aws_instance":                    "id",
-					"azurerm_linux_virtual_machine":   "id",
-					"azurerm_windows_virtual_machine": "id",
-				}[resource.Type]
-				if !isValid {
-					// The resource type doesn't support
-					// automatically setting the instance ID.
-					continue
-				}
-				instanceIDRaw, valid := resource.AttributeValues[key]
-				if !valid {
-					continue
-				}
-				instanceID, valid := instanceIDRaw.(string)
-				if !valid {
-					continue
-				}
-				agent.Auth = &proto.Agent_InstanceId{
-					InstanceId: instanceID,
-				}
-			}
-
-			resources = append(resources, &proto.Resource{
-				Name:   resource.Name,
-				Type:   resource.Type,
-				Agents: resourceAgents,
-			})
+			return nil, err
 		}
 	}
 
@@ -621,10 +376,10 @@ func parseTerraformApply(ctx context.Context, terraform *tfexec.Terraform, state
 	}, nil
 }
 
-// getTerraformState pulls and merges any remote terraform state into the given
+// pullTerraformState pulls and merges any remote terraform state into the given
 // path and reads the merged state. If there is no state, `os.ErrNotExist` will
 // be returned.
-func getTerraformState(ctx context.Context, terraform *tfexec.Terraform, statefilePath string) (*tfjson.State, error) {
+func pullTerraformState(ctx context.Context, terraform *tfexec.Terraform, statefilePath string) (*tfjson.State, error) {
 	statefile, err := os.OpenFile(statefilePath, os.O_CREATE|os.O_RDWR, 0600)
 	if err != nil {
 		return nil, xerrors.Errorf("open statefile %q: %w", statefilePath, err)
@@ -680,76 +435,4 @@ func convertTerraformLogLevel(logLevel string) (proto.LogLevel, error) {
 	default:
 		return proto.LogLevel(0), xerrors.Errorf("invalid log level %q", logLevel)
 	}
-}
-
-// findDirectDependencies maps Terraform resources to their children nodes.
-// This parses GraphViz output from Terraform which isn't ideal, but seems reliable.
-func findDirectDependencies(rawGraph string) (map[string][]string, error) {
-	parsedGraph, err := gographviz.ParseString(rawGraph)
-	if err != nil {
-		return nil, xerrors.Errorf("parse graph: %w", err)
-	}
-	graph, err := gographviz.NewAnalysedGraph(parsedGraph)
-	if err != nil {
-		return nil, xerrors.Errorf("analyze graph: %w", err)
-	}
-	direct := map[string][]string{}
-	for _, node := range graph.Nodes.Nodes {
-		label, exists := node.Attrs["label"]
-		if !exists {
-			continue
-		}
-		label = strings.Trim(label, `"`)
-		direct[label] = findDependenciesWithLabels(graph, node.Name)
-	}
-
-	return direct, nil
-}
-
-// findDependenciesWithLabels recursively finds nodes with labels (resource and data nodes)
-// to build a dependency tree.
-func findDependenciesWithLabels(graph *gographviz.Graph, nodeName string) []string {
-	dependencies := make([]string, 0)
-	for destination := range graph.Edges.SrcToDsts[nodeName] {
-		dependencyNode, exists := graph.Nodes.Lookup[destination]
-		if !exists {
-			continue
-		}
-		label, exists := dependencyNode.Attrs["label"]
-		if !exists {
-			dependencies = append(dependencies, findDependenciesWithLabels(graph, dependencyNode.Name)...)
-			continue
-		}
-		label = strings.Trim(label, `"`)
-		dependencies = append(dependencies, label)
-	}
-	return dependencies
-}
-
-// findAgents recursively searches through resource dependencies
-// to find associated agents. Nested is required for indirect
-// dependency matching.
-func findAgents(resourceDependencies map[string][]string, agents map[string]*proto.Agent, resourceLabel string) []*proto.Agent {
-	resourceNode, exists := resourceDependencies[resourceLabel]
-	if !exists {
-		return []*proto.Agent{}
-	}
-	// Associate resources that depend on an agent.
-	resourceAgents := make([]*proto.Agent, 0)
-	for _, dep := range resourceNode {
-		var has bool
-		agent, has := agents[dep]
-		if !has {
-			resourceAgents = append(resourceAgents, findAgents(resourceDependencies, agents, dep)...)
-			continue
-		}
-		resourceAgents = append(resourceAgents, agent)
-	}
-	return resourceAgents
-}
-
-// convertAddressToLabel returns the Terraform address without the count
-// specifier. eg. "module.ec2_dev.ec2_instance.dev[0]" becomes "module.ec2_dev.ec2_instance.dev"
-func convertAddressToLabel(address string) string {
-	return strings.Split(address, "[")[0]
 }
