+++ provisioner/terraform/provision_test.go
@@ -121,369 +121,6 @@ provider "coder" {
 			"main.tf": `a`,
 		},
 		Error: true,
-	}, {
-		Name: "dryrun-single-resource",
-		Files: map[string]string{
-			"main.tf": `resource "null_resource" "A" {}`,
-		},
-		Request: &proto.Provision_Request{
-			Type: &proto.Provision_Request_Start{
-				Start: &proto.Provision_Start{
-					DryRun: true,
-				},
-			},
-		},
-		Response: &proto.Provision_Response{
-			Type: &proto.Provision_Response_Complete{
-				Complete: &proto.Provision_Complete{
-					Resources: []*proto.Resource{{
-						Name: "A",
-						Type: "null_resource",
-					}},
-				},
-			},
-		},
-	}, {
-		Name: "dryrun-conditional-single-resource",
-		Files: map[string]string{
-			"main.tf": `
-			variable "test" {
-				default = "no"
-			}
-			resource "null_resource" "A" {
-				count = var.test == "yes" ? 1 : 0
-			}`,
-		},
-		Response: &proto.Provision_Response{
-			Type: &proto.Provision_Response_Complete{
-				Complete: &proto.Provision_Complete{
-					Resources: nil,
-				},
-			},
-		},
-	}, {
-		Name: "resource-associated-with-agent",
-		Files: map[string]string{
-			"main.tf": provider + `
-			resource "coder_agent" "A" {
-				os = "windows"
-				arch = "arm64"
-				dir = "C:\\System32"
-			}
-			resource "null_resource" "A" {
-				depends_on = [
-					coder_agent.A
-				]
-			}`,
-		},
-		Request: &proto.Provision_Request{
-			Type: &proto.Provision_Request_Start{
-				Start: &proto.Provision_Start{
-					Metadata: &proto.Provision_Metadata{},
-				},
-			},
-		},
-		Response: &proto.Provision_Response{
-			Type: &proto.Provision_Response_Complete{
-				Complete: &proto.Provision_Complete{
-					Resources: []*proto.Resource{{
-						Name: "A",
-						Type: "null_resource",
-						Agents: []*proto.Agent{{
-							Name:            "A",
-							OperatingSystem: "windows",
-							Architecture:    "arm64",
-							Directory:       "C:\\System32",
-							Auth: &proto.Agent_Token{
-								Token: "",
-							},
-						}},
-					}},
-				},
-			},
-		},
-	}, {
-		Name: "dryrun-resource-associated-with-agent",
-		Files: map[string]string{
-			"main.tf": provider + `
-			data "coder_workspace" "me" {}
-			resource "coder_agent" "A" {
-				count = 1
-				os = "linux"
-				arch = "amd64"
-				env = {
-					test: "example"
-				}
-				startup_script = "code-server"
-			}
-			resource "null_resource" "A" {
-				depends_on = [
-					coder_agent.A[0]
-				]
-				count = data.coder_workspace.me.start_count
-			}`,
-		},
-		Request: &proto.Provision_Request{
-			Type: &proto.Provision_Request_Start{
-				Start: &proto.Provision_Start{
-					DryRun:   true,
-					Metadata: &proto.Provision_Metadata{},
-				},
-			},
-		},
-		Response: &proto.Provision_Response{
-			Type: &proto.Provision_Response_Complete{
-				Complete: &proto.Provision_Complete{
-					Resources: []*proto.Resource{{
-						Name: "A",
-						Type: "null_resource",
-						Agents: []*proto.Agent{{
-							Name:            "A",
-							OperatingSystem: "linux",
-							Architecture:    "amd64",
-							Auth:            &proto.Agent_Token{},
-							Env: map[string]string{
-								"test": "example",
-							},
-							StartupScript: "code-server",
-						}},
-					}},
-				},
-			},
-		},
-	}, {
-		Name: "resource-manually-associated-with-agent",
-		Files: map[string]string{
-			"main.tf": provider + `
-			resource "coder_agent" "A" {
-				os = "darwin"
-				arch = "amd64"
-			}
-			resource "null_resource" "A" {
-				depends_on = [
-					coder_agent.A
-				]
-			}
-			resource "coder_agent_instance" "A" {
-				agent_id = coder_agent.A.id
-				instance_id = "bananas"
-			}
-			`,
-		},
-		Request: &proto.Provision_Request{
-			Type: &proto.Provision_Request_Start{
-				Start: &proto.Provision_Start{
-					Metadata: &proto.Provision_Metadata{},
-				},
-			},
-		},
-		Response: &proto.Provision_Response{
-			Type: &proto.Provision_Response_Complete{
-				Complete: &proto.Provision_Complete{
-					Resources: []*proto.Resource{{
-						Name: "A",
-						Type: "null_resource",
-						Agents: []*proto.Agent{{
-							Name:            "A",
-							OperatingSystem: "darwin",
-							Architecture:    "amd64",
-							Auth: &proto.Agent_InstanceId{
-								InstanceId: "bananas",
-							},
-						}},
-					}},
-				},
-			},
-		},
-	}, {
-		Name: "resource-manually-associated-with-multiple-agents",
-		Files: map[string]string{
-			"main.tf": provider + `
-			resource "coder_agent" "A" {
-				os = "darwin"
-				arch = "amd64"
-			}
-			resource "coder_agent" "B" {
-				os = "linux"
-				arch = "amd64"
-			}
-			resource "null_resource" "A" {
-				depends_on = [
-					coder_agent.A,
-					coder_agent.B
-				]
-			}
-			resource "coder_agent_instance" "A" {
-				agent_id = coder_agent.A.id
-				instance_id = "bananas"
-			}
-			`,
-		},
-		Request: &proto.Provision_Request{
-			Type: &proto.Provision_Request_Start{
-				Start: &proto.Provision_Start{
-					Metadata: &proto.Provision_Metadata{},
-				},
-			},
-		},
-		Response: &proto.Provision_Response{
-			Type: &proto.Provision_Response_Complete{
-				Complete: &proto.Provision_Complete{
-					Resources: []*proto.Resource{{
-						Name: "A",
-						Type: "null_resource",
-						Agents: []*proto.Agent{{
-							Name:            "A",
-							OperatingSystem: "darwin",
-							Architecture:    "amd64",
-							Auth: &proto.Agent_InstanceId{
-								InstanceId: "bananas",
-							},
-						}, {
-							Name:            "B",
-							OperatingSystem: "linux",
-							Architecture:    "amd64",
-							Auth: &proto.Agent_Token{
-								Token: "",
-							},
-						}},
-					}},
-				},
-			},
-		},
-	}, {
-		Name: "dryrun-resource-separated-from-agent",
-		Files: map[string]string{
-			"main.tf": provider + `
-			resource "coder_agent" "A" {
-				os = "darwin"
-				arch = "amd64"
-			}
-			data "null_data_source" "values" {
-				inputs = {
-					script = coder_agent.A.init_script
-				}
-			}
-			resource "null_resource" "A" {
-				depends_on = [
-					data.null_data_source.values
-				]
-			}
-			`,
-		},
-		Request: &proto.Provision_Request{
-			Type: &proto.Provision_Request_Start{
-				Start: &proto.Provision_Start{
-					Metadata: &proto.Provision_Metadata{},
-					DryRun:   true,
-				},
-			},
-		},
-		Response: &proto.Provision_Response{
-			Type: &proto.Provision_Response_Complete{
-				Complete: &proto.Provision_Complete{
-					Resources: []*proto.Resource{{
-						Name: "A",
-						Type: "null_resource",
-						Agents: []*proto.Agent{{
-							Name:            "A",
-							OperatingSystem: "darwin",
-							Architecture:    "amd64",
-							Auth:            &proto.Agent_Token{},
-						}},
-					}},
-				},
-			},
-		},
-	}, {
-		Name: "resource-separated-from-agent",
-		Files: map[string]string{
-			"main.tf": provider + `
-			resource "coder_agent" "A" {
-				os = "darwin"
-				arch = "amd64"
-			}
-			data "null_data_source" "values" {
-				inputs = {
-					script = coder_agent.A.init_script
-				}
-			}
-			resource "null_resource" "A" {
-				depends_on = [
-					data.null_data_source.values
-				]
-			}
-			`,
-		},
-		Request: &proto.Provision_Request{
-			Type: &proto.Provision_Request_Start{
-				Start: &proto.Provision_Start{
-					Metadata: &proto.Provision_Metadata{},
-				},
-			},
-		},
-		Response: &proto.Provision_Response{
-			Type: &proto.Provision_Response_Complete{
-				Complete: &proto.Provision_Complete{
-					Resources: []*proto.Resource{{
-						Name: "A",
-						Type: "null_resource",
-						Agents: []*proto.Agent{{
-							Name:            "A",
-							OperatingSystem: "darwin",
-							Architecture:    "amd64",
-							Auth:            &proto.Agent_Token{},
-						}},
-					}},
-				},
-			},
-		},
-	}, {
-		Name: "agent-with-app",
-		Files: map[string]string{
-			"main.tf": provider + `
-			resource "coder_agent" "A" {
-				os = "darwin"
-				arch = "amd64"
-			}
-			resource "null_resource" "A" {
-				depends_on = [
-					coder_agent.A
-				]
-			}
-			resource "coder_app" "A" {
-				agent_id = coder_agent.A.id
-				command = "vim"
-			}
-			`,
-		},
-		Request: &proto.Provision_Request{
-			Type: &proto.Provision_Request_Start{
-				Start: &proto.Provision_Start{
-					Metadata: &proto.Provision_Metadata{},
-				},
-			},
-		},
-		Response: &proto.Provision_Response{
-			Type: &proto.Provision_Response_Complete{
-				Complete: &proto.Provision_Complete{
-					Resources: []*proto.Resource{{
-						Name: "A",
-						Type: "null_resource",
-						Agents: []*proto.Agent{{
-							Name:            "A",
-							OperatingSystem: "darwin",
-							Architecture:    "amd64",
-							Auth:            &proto.Agent_Token{},
-							Apps: []*proto.App{{
-								Name:    "A",
-								Command: "vim",
-							}},
-						}},
-					}},
-				},
-			},
-		},
 	}} {
 		testCase := testCase
 		t.Run(testCase.Name, func(t *testing.T) {
