+++ provisioner/terraform/testdata/calling-module/module/module.tf
@@ -1,0 +1,9 @@
+variable "script" {
+    type = string
+}
+
+resource "null_resource" "example" {
+    depends_on = [
+      var.script
+    ]
+}
